backend/requirements.txt
Flask
Flask-Cors
asyncio
flask_caching
google-genai